Added a better assigmment of letters to groups
I wanted to take advantage of my work on classification success rates
(Cats, Dogs, and the Speed of Light) to better help me assign letters
to groups. I also wanted to determine how accurate the classification
would be, so I made the computations to do so. The results were pretty
gratifying.